Test: Extract common verifier output assertions helper functionality
